state mutation goes to state, and un-nest Builder.Util
