+++ lib/grpc_reflection/service/state.ex
@@ -4,12 +4,41 @@ defmodule GrpcReflection.Service.State do
   defstruct services: [], files: %{}, symbols: %{}, extensions: %{}
 
   @type descriptor_t :: GrpcReflection.Server.descriptor_t()
+  @type entry_t :: %{optional(binary()) => descriptor_t()}
+
   @type t :: %__MODULE__{
           services: list(module()),
-          files: %{optional(binary()) => descriptor_t()},
-          symbols: %{optional(binary()) => descriptor_t()},
+          files: entry_t(),
+          symbols: entry_t(),
           extensions: %{optional(binary()) => list(integer())}
         }
+
+  @spec new(list(module)) :: t()
+  def new(services \\ []), do: %__MODULE__{services: services}
+
+  @spec merge(t(), t()) :: t()
+  def merge(state1, state2) do
+    %__MODULE__{
+      services: Enum.uniq(state1.services ++ state2.services),
+      files: Map.merge(state1.files, state2.files),
+      symbols: Map.merge(state1.symbols, state2.symbols),
+      extensions: Map.merge(state1.extensions, state2.extensions)
+    }
+  end
+
+  @spec add_files(t(), entry_t()) :: t()
+  def add_files(%__MODULE__{} = state, files) do
+    %{state | files: Map.merge(files, state.files)}
+  end
+
+  @spec add_symbols(t(), entry_t()) :: t()
+  def add_symbols(%__MODULE__{} = state, symbols) do
+    %{state | symbols: Map.merge(symbols, state.symbols)}
+  end
+
+  def add_extensions(%__MODULE__{} = state, extensions) do
+    %{state | extensions: Map.merge(extensions, state.extensions)}
+  end
 
   def lookup_services(%__MODULE__{services: services}) do
     Enum.map(services, fn service_mod -> service_mod.__meta__(:name) end)
